Update fields types HAJ-1

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/ApplicantContactDto.java
@@ -27,7 +27,7 @@ public class ApplicantContactDto implements Serializable {
     private String languageList;
     private String email;
     private int localMobileNumber;
-    private int intlMobileNumber;
+    private long intlMobileNumber;
     private String countryCode;
     private String streetName;
     private String districtName;
